Huge Overhaul
Too much to List. Had to rewrite a lot of stuff as I discovered bug after bug after bug, all from my lazy work earlier....

I'm in the process of rewriting parseAndPlot(), but recursion is crashing the program, so I'll have to find a different route. Leave the notes though.